requirements.txt
altgraph==0.17.4
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorzero==2.0
Flask==3.1.0
Flask-SocketIO==5.5.1
future==1.0.0
gevent==25.4.2
gpiozero==2.0.1
greenlet==3.2.1
h11==0.16.0
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lgpio==0.2.2.0
MarkupSafe==3.0.2
packaging==25.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyserial==3.5
python-engineio==4.12.0
python-socketio==5.13.0
PyYAML==6.0.2
requests==2.32.3
rpi-lgpio==0.6
simple-websocket==1.1.0
six==1.17.0
socketIO-client==0.7.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
